Added the old programs into the old folder
These are some of the other programs that I made as I was working on the project. These didn't end up being fully used in the final program, or maybe were just for testing something or doing analysis. The outlier is CountCodes, where I tried to package up the main algorithm that I had at the time to something like fully usable, but I didn't really enjoy the process and it really complicated the code, and I found a better algorithm anyway, so it's not very useful anymore.